+++ src/components/entries/NumberField.js
@@ -139,7 +139,7 @@ export default function NumberFieldEntry(props) {
       newValidationError = validate(newValue) || null;
     }
 
-    setValue(newValue);
+    setValue(newValue, newValidationError);
     setLocalError(newValidationError);
   };